Split the resolution setting into Window Size and Rendering Resolution (internal resolution).
Gets rid of our "antialiasing" (simply crank up internal resolution to 2x your window size).

Proper MSAA is coming later.

+++ UI/CwCheatScreen.cpp
@@ -49,7 +49,7 @@ std::vector<std::string> CwCheatScreen::CreateCodeList() {
 		if (cheatList[i].substr(0, 3) == "_C1") {
 			formattedList.push_back(cheatList[i].substr(4));
 			enableCheat[j++] = true;
-			locations.push_back(i);
+			locations.push_back((int)i);
 		}
 		if (cheatList[i].substr(0, 3) == "_C0") {
 			formattedList.push_back(cheatList[i].substr(4));
